LayoutTests/accessibility/mac/meter-gauge-value-description-expected.txt


This tests the gauge value description for meter elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS meter1.valueDescription is 'AXValueDescription: 1'
PASS meter2.valueDescription is 'AXValueDescription: 5, suboptimal value'
PASS meter3.valueDescription is 'AXValueDescription: 50, optimal value'
PASS meter4.valueDescription is 'AXValueDescription: 90, critical value'
PASS meter5.valueDescription is 'AXValueDescription: 50 hours, optimal value'
PASS successfullyParsed is true

TEST COMPLETE
